pyproject.toml
[tool.poetry]
name = "crypto-wallet"
description = "Crypto Wallet"
version = "0.1.0"
readme = "README.md"
authors = ["crypto_wallet.nmelis.com"]


[tool.poetry.dependencies]
python = "3.9.10"

django = { version = "^3.2", extras = ["argon2"] }
django-split-settings = "^1.1"
django-axes = "^5.33"
django-csp = "^3.7"
django-health-check = "^3.16"
django-http-referrer-policy = "^1.1"
django-permissions-policy = "^4.10"
django-stubs-ext = "^0.4"

psycopg2-binary = "^2.9"
gunicorn = "^20.0"
python-decouple = "^3.6"
structlog = "^21.5"
djangorestframework = "^3.13.1"
celery = "5.2.6"
redis = "^4.3.1"


[tool.poetry.dev-dependencies]
django-debug-toolbar = "^3.4"
django-querycount = "^0.7"
django-migration-linter = "^4.1"
django-extra-checks = "^0.11"
django-coverage-plugin = "^2.0"
nplusone = "^1.0"

wemake-python-styleguide = "^0.16"
flake8-pytest-style = "^1.6"
flake8-django = "^1.1"
flake8-logging-format = "^0.6"
nitpick = "^0.32"

pytest = "^7.1"
pytest-django = "^4.5"
pytest-cov = "^3.0"
pytest-randomly = "^3.12"
pytest-deadfixtures = "^2.2"
pytest-testmon = "^1.3"
pytest-timeout = "^2.1"
django-test-migrations = "^1.2"
hypothesis = "^6.46"

mypy = "^0.942"
django-stubs = "^1.10"

sphinx = "^4.5"
sphinx-autodoc-typehints = "^1.18"
tomlkit = "^0.10"
doc8 = "^0.11"

yamllint = "^1.26"
safety = "^1.10"
dotenv-linter = "^0.3"
polint = "^0.4"
dennis = "^0.9"
dump-env = "^1.3"
ipython = "^8.3"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.nitpick]
style = "https://raw.githubusercontent.com/wemake-services/wemake-python-styleguide/master/styles/nitpick-style-wemake.toml"
